feat(calendar): implement CalendarView component for improved event display
- Created a new CalendarView component to encapsulate calendar rendering logic
- Moved event handling and navigation functionality to CalendarView
- Refactored event filtering logic into a utility function
- Updated CalendarPage to utilize the new CalendarView component

Files modified:
- src/app/dashboard/calendar/page.jsx
- src/components/dashboard/calendar/CalendarView.jsx
- src/lib/utils.js
